+++ src/app/loadout/loadout-ui/Sockets.m.scss.d.ts
@@ -2,6 +2,7 @@
 // Please do not change this file!
 interface CssExports {
   'lockedItems': string;
+  'small': string;
 }
 export const cssExports: CssExports;
 export default cssExports;