+++ src/app/shell/filters.ts
@@ -112,10 +112,8 @@ export const acquisitionRecencyComparator = reverseComparator(
 const ITEM_COMPARATORS: { [key: string]: Comparator<DimItem> } = {
   typeName: compareBy((item: DimItem) => item.typeName),
   rarity: compareBy(rarity),
-  primStat: reverseComparator(compareBy((item: DimItem) => item.primStat?.value ?? 0)),
-  basePower: reverseComparator(
-    compareBy((item: DimItem) => item.basePower || item.primStat?.value)
-  ),
+  primStat: reverseComparator(compareBy((item: DimItem) => item.primaryStat?.value ?? 0)),
+  basePower: reverseComparator(compareBy((item: DimItem) => item.power)),
   // This only sorts by D1 item quality
   rating: reverseComparator(
     compareBy((item: DimItem & { quality: { min: number } }) => {