+++ docs/CHANGELOG.md
@@ -1,6 +1,18 @@
 ## Next
 
 * If a loadout has items for multiple character classes in it, applying it to a character behaves as if only the items that can be equipped on that character are in the loadout.
+* Fixed an issue where clicking a mod icon in the Loadout Optimizer would select more than one of the mod.
+
+## 6.89.0 <span class="changelog-date">(2021-10-31)</span>
+
+## 6.88.1 <span class="changelog-date">(2021-10-28)</span>
+
+* `modslot:activity` now identifies Armor 2.0 items that have a modslot related to an activity (currently, a raid or a Nightmare mod slot).
+* Fix an issue where an invalid query is passed to the Loadout Optimizer when you click a mod socket.
+
+### Beta Only
+
+* Loadouts can now show you an assignment strategy for mods. It optimizes for the least number of unassigned mods.
 
 ## 6.88.0 <span class="changelog-date">(2021-10-24)</span>